+++ packages/components/releaseNotes/components.md
@@ -4,7 +4,13 @@ Components, models, actions, and utility functions for LabKey applications and p
 ### version 3.X
 *Released*: X January 2024
 - Support Date-only or Time-only fields
-  - TODO
+  - add container.format.timeFormat
+  - add date and time fields as allowed domain field types
+  - update ConfirmDataTypeChangeModal to allow date/time type convert
+  - modify DatePickerInput/DateInputCell to support time-only fields
+  - wire up time-only fields for QueryFormInputs, DetailDisplay, FilterExpressionView and editable grid
+  - add util for parsing/formatting time only fields
+
 
 ### version 3.11.0
 *Released*: 29 January 2024
