+++ packages/components/src/internal/components/domainproperties/ConfirmDataTypeChangeModal.tsx
@@ -24,9 +24,8 @@ export const ConfirmDataTypeChangeModal: FC<Props> = memo(props => {
     const origTypeLabel = getDataTypeConfirmDisplayText(originalRangeURI);
     const newTypeLabel = getDataTypeConfirmDisplayText(newDataType.rangeURI);
 
-    const reversible =
-        (originalRangeURI === DATE_RANGE_URI && newDataType.rangeURI === DATETIME_RANGE_URI) ||
-        (originalRangeURI === DATETIME_RANGE_URI && newDataType.rangeURI === DATE_RANGE_URI);
+    const reversible = (PropDescType.isDate(originalRangeURI) && PropDescType.isDateTime(newDataType.rangeURI)) ||
+        (PropDescType.isDateTime(originalRangeURI) && PropDescType.isDate(newDataType.rangeURI));
 
     let dataLossWarning = null;
     if (
